pyproject.toml
[tool.poetry]
name = "pinecone-demo"
version = "0.1.0"
description = ""
authors = ["Nir Gazit <nir.gzt@gmail.com>"]
readme = "README.md"
packages = [{include = "pinecone_demo"}]

[tool.poetry.dependencies]
python = "~3.11"
openai = "^1.9.0"
pinecone-client = {extras = ["grpc"], version = "^3.2.2"}
pinecone-datasets = "^0.7.0"
langchain = "^0.1.17"
tiktoken = "^0.6.0"
traceloop-sdk = "^0.18.2"
langchain-community = "^0.0.37"
langchain-pinecone = "^0.1.0"
langchain-openai = "^0.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
